+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to
 
 ## [Unreleased]
 
+## [2.4.0] - 2023-11-13
+
 ### Added
 
 - New [`deep_intersect/2`][deep_intersect-2] function
@@ -94,7 +96,8 @@ and this project adheres to
 
 Initial release
 
-[unreleased]: https://github.com/eproxus/mapz/compare/v2.3.0...HEAD
+[unreleased]: https://github.com/eproxus/mapz/compare/v2.4.0...HEAD
+[2.4.0]: https://github.com/eproxus/mapz/compare/v2.3.0...v2.4.0
 [2.3.0]: https://github.com/eproxus/mapz/compare/v2.2.1...v2.3.0
 [2.2.0]: https://github.com/eproxus/mapz/compare/v2.1.1...v2.2.0
 [2.1.1]: https://github.com/eproxus/mapz/compare/v2.1.0...v2.1.1